backend/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel", "pybind11>=2.11.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-engine"
version = "0.7.0"
description = "High-performance trading engine with Python orchestration layer"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Nexus Team", email = "noreply@nexus-engine.com"}
]
keywords = ["trading", "backtest", "algorithmic-trading", "high-performance", "cpp", "python"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: C++",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "pybind11>=2.11.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "finnhub-python>=2.4.0",
    "websocket-client>=1.6.0",
    "alpha-vantage>=2.3.0",
    "nasdaq-data-link>=1.0.0",
    "fredapi>=0.5.0",
    "prometheus-client>=0.18.0",
    "python-json-logger>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["nexus_bindings", "domain", "application", "infrastructure"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src --cov-report=html --cov-report=term"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []